Fix broadcasting
Due to react-broadcasts generic implementation, nested ThemeProviders
would not merge their themes, they would override them. Fixed that
behaviour, everything now works as expected :tada:
